+++ lib/freshdesk-rest.rb
@@ -3,6 +3,10 @@ require_relative 'freshdesk-rest/version'
 require_relative 'freshdesk-rest/configuration'
 require_relative 'freshdesk-rest/resource/contact'
 require_relative 'freshdesk-rest/resource/ticket'
+require_relative 'freshdesk-rest/resource/solutions'
+require_relative 'freshdesk-rest/resource/solutions/article'
+require_relative 'freshdesk-rest/resource/solutions/category'
+require_relative 'freshdesk-rest/resource/solutions/folder'
 
 module Freshdesk
   module Rest